Add selectable option into Menu documentation
I had to read the source code to figure out that there was a property called selectable. So hopefully this will help save someone time. (Though this is an option in the rc-menu I made a PR there to add it in the doc also https://github.com/react-component/menu/pull/79)

(I was trying to make a contextmenu out of this menu component which is why I needed to turn selectable off)